requirements.txt
Django
django-cors-headers
djangorestframework
django-filter
djangorestframework-simplejwt
django-simple-history
psycopg2